Revert "[refactor] 트랜잭션 범위 분리 (#46)"
This reverts commit c6e1d4cc72a0ff34e6474930d707486c1fdf9dfc.
